+++ xclim/indices/_threshold.py
@@ -1,4 +1,5 @@
 # noqa: D100
+import warnings
 from typing import Optional
 
 import numpy as np
@@ -48,8 +49,13 @@ __all__ = [
     "hot_spell_frequency",
     "hot_spell_max_length",
     "snow_cover_duration",
+    "tn_days_above",
     "tn_days_below",
+    "tg_days_above",
+    "tg_days_below",
     "tx_days_above",
+    "tx_days_below",
+    "tropical_nights",
     "warm_day_frequency",
     "warm_night_frequency",
     "wetdays",
@@ -62,7 +68,6 @@ __all__ = [
     "maximum_consecutive_wet_days",
     "sea_ice_area",
     "sea_ice_extent",
-    "tropical_nights",
 ]
 
 
@@ -1073,7 +1078,7 @@ def hot_spell_max_length(
     Returns
     -------
     xarray.DataArray, [time]
-      Maximum length of continuous hot days at the wanted frequency
+      Maximum length of continuous hot days at the wanted frequency.
 
     Notes
     -----
@@ -1187,6 +1192,42 @@ def snow_cover_duration(
 
 
 @declare_units(tasmin="[temperature]", thresh="[temperature]")
+def tn_days_above(
+    tasmin: xarray.DataArray, thresh: str = "20.0 degC", freq: str = "YS"
+):  # noqa: D401
+    """Number of days with tasmin above a threshold (number of tropical nights).
+
+    Number of days where daily minimum temperature exceeds a threshold.
+
+    Parameters
+    ----------
+    tasmin : xarray.DataArray
+      Minimum daily temperature.
+    thresh : str
+      Threshold temperature on which to base evaluation.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [time]
+      Number of days where tasmin > threshold.
+
+    Notes
+    -----
+    Let :math:`TN_{ij}` be the daily minimum temperature at day :math:`i` of period :math:`j`. Then
+    counted is the number of days where:
+
+    .. math::
+
+        TN_{ij} > Threshold [℃]
+    """
+    thresh = convert_units_to(thresh, tasmin)
+    f = threshold_count(tasmin, ">", thresh, freq)
+    return to_agg_units(f, tasmin, "count")
+
+
+@declare_units(tasmin="[temperature]", thresh="[temperature]")
 def tn_days_below(
     tasmin: xarray.DataArray, thresh: str = "-10.0 degC", freq: str = "YS"
 ) -> xarray.DataArray:  # noqa: D401
@@ -1206,7 +1247,7 @@ def tn_days_below(
     Returns
     -------
     xarray.DataArray, [time]
-      The number of days with tasmin < threshold per period.
+      Number of days where tasmin < threshold.
 
     Notes
     -----
@@ -1215,20 +1256,92 @@ def tn_days_below(
 
     .. math::
 
-        TX_{ij} < Threshold [℃]
+        TN_{ij} < Threshold [℃]
     """
     thresh = convert_units_to(thresh, tasmin)
     f1 = threshold_count(tasmin, "<", thresh, freq)
     return to_agg_units(f1, tasmin, "count")
 
 
+@declare_units(tas="[temperature]", thresh="[temperature]")
+def tg_days_above(
+    tas: xarray.DataArray, thresh: str = "10.0 degC", freq: str = "YS"
+):  # noqa: D401
+    """Number of days with tas above a threshold.
+
+    Number of days where daily mean temperature exceeds a threshold.
+
+    Parameters
+    ----------
+    tas : xarray.DataArray
+      Mean daily temperature.
+    thresh : str
+      Threshold temperature on which to base evaluation.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [time]
+      Number of days where tas > threshold.
+
+    Notes
+    -----
+    Let :math:`TG_{ij}` be the daily mean temperature at day :math:`i` of period :math:`j`. Then
+    counted is the number of days where:
+
+    .. math::
+
+        TG_{ij} > Threshold [℃]
+    """
+    thresh = convert_units_to(thresh, tas)
+    f = threshold_count(tas, ">", thresh, freq)
+    return to_agg_units(f, tas, "count")
+
+
+@declare_units(tas="[temperature]", thresh="[temperature]")
+def tg_days_below(
+    tas: xarray.DataArray, thresh: str = "10.0 degC", freq: str = "YS"
+):  # noqa: D401
+    """Number of days with tas below a threshold.
+
+    Number of days where daily mean temperature is below a threshold.
+
+    Parameters
+    ----------
+    tas : xarray.DataArray
+      Mean daily temperature.
+    thresh : str
+      Threshold temperature on which to base evaluation.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [time]
+      Number of days where tas < threshold.
+
+    Notes
+    -----
+    Let :math:`TG_{ij}` be the daily mean temperature at day :math:`i` of period :math:`j`. Then
+    counted is the number of days where:
+
+    .. math::
+
+        TG_{ij} < Threshold [℃]
+    """
+    thresh = convert_units_to(thresh, tas)
+    f1 = threshold_count(tas, "<", thresh, freq)
+    return to_agg_units(f1, tas, "count")
+
+
 @declare_units(tasmax="[temperature]", thresh="[temperature]")
 def tx_days_above(
     tasmax: xarray.DataArray, thresh: str = "25.0 degC", freq: str = "YS"
 ) -> xarray.DataArray:  # noqa: D401
-    """Number of summer days.
+    """Number of days with tasmax above a threshold (number of summer days).
 
-    Number of days where daily maximum temperature exceed a threshold.
+    Number of days where daily maximum temperature exceeds a threshold.
 
     Parameters
     ----------
@@ -1242,7 +1355,7 @@ def tx_days_above(
     Returns
     -------
     xarray.DataArray, [time]
-      Number of summer days (number of days with tasmax > threshold per period).
+      Number of days where tasmax > threshold (number of summer days).
 
     Notes
     -----
@@ -1256,6 +1369,42 @@ def tx_days_above(
     thresh = convert_units_to(thresh, tasmax)
     f = threshold_count(tasmax, ">", thresh, freq)
     return to_agg_units(f, tasmax, "count")
+
+
+@declare_units(tasmax="[temperature]", thresh="[temperature]")
+def tx_days_below(
+    tasmax: xarray.DataArray, thresh: str = "25.0 degC", freq: str = "YS"
+):  # noqa: D401
+    """Number of days with tmax below a threshold.
+
+    Number of days where daily maximum temperature is below a threshold.
+
+    Parameters
+    ----------
+    tasmax : xarray.DataArray
+      Maximum daily temperature.
+    thresh : str
+      Threshold temperature on which to base evaluation.
+    freq : str
+      Resampling frequency.
+
+    Returns
+    -------
+    xarray.DataArray, [time]
+      Number of days where tasmin < threshold.
+
+    Notes
+    -----
+    Let :math:`TN_{ij}` be the daily minimum temperature at day :math:`i` of period :math:`j`. Then
+    counted is the number of days where:
+
+    .. math::
+
+        TN_{ij} < Threshold [℃]
+    """
+    thresh = convert_units_to(thresh, tasmax)
+    f1 = threshold_count(tasmax, "<", thresh, freq)
+    return to_agg_units(f1, tasmax, "count")
 
 
 @declare_units(tasmax="[temperature]", thresh="[temperature]")
@@ -1495,7 +1644,7 @@ def maximum_consecutive_frost_free_days(
 def maximum_consecutive_tx_days(
     tasmax: xarray.DataArray, thresh: str = "25 degC", freq: str = "YS"
 ) -> xarray.DataArray:
-    r"""Maximum number of consecutive summer days.
+    r"""Maximum number of consecutive days with tasmax above a threshold (summer days).
 
     Return the maximum number of consecutive days within the period where
     the maximum temperature is above a certain threshold.
@@ -1512,7 +1661,7 @@ def maximum_consecutive_tx_days(
     Returns
     -------
     xarray.DataArray, [time]
-      The maximum number of consecutive summer days (tasmax > threshold per period).
+      The maximum number of days with tasmax > thresh per periods (summer days).
 
     Notes
     -----
@@ -1644,10 +1793,20 @@ def tropical_nights(
     .. math::
 
         TN_{ij} > Threshold [℃]
+
+    Warnings
+    --------
+    The `tropical_nights` indicator is being deprecated in favour of `tn_days_above` with `thresh="20 degC" by default.
+    This change will be effectuated in a future version of xclim.
     """
-    thresh = convert_units_to(thresh, tasmin)
-    out = threshold_count(tasmin, ">", thresh, freq)
-    return to_agg_units(out, tasmin, "count")
+    warnings.warn(
+        "The `tropical_nights` indice is being deprecated in favour of `tn_days_above` with `thresh='20 degC'. "
+        "This indice will be removed in `xclim>=0.28.0`. Please update your scripts accordingly.",
+        UserWarning,
+        stacklevel=3,
+    )
+
+    return tn_days_above(tasmin, thresh=thresh, freq=freq)
 
 
 @declare_units(tas="[temperature]", thresh="[temperature]", sum_thresh="K days")